Refactor project structure and update dependencies
- Renamed project from "painting-ai" to "painting" in package.json and manifest.json.
- Updated dependency versions in package.json for better stability.
- Integrated Capacitor Filesystem to save canvas images directly to device storage.
- Improved save functionality in Canvas component to handle both native and browser environments.
- Changed ColorPicker and SaveFab button colors from primary to secondary for better UI consistency.
- Removed unused ImageAgent, ProblemSolver, and RootAgent logic files to clean up the codebase.
- Removed AI processing logic from HomePage component.
- Added a comprehensive privacy policy document.
- Created MainActivity.java for Android integration.
