+++ PRIVACY_POLICY.md
@@ -1,0 +1,28 @@
+# Privacy Policy
+
+_Last updated: May 2025_
+
+This privacy policy describes how the "painting" app handles your information.
+
+## Information Collection
+
+- The app does **not** collect, store, or transmit any personal information.
+- Drawings you create are processed and saved **locally** on your device.
+- No user data is sent to any server or third party.
+
+## Permissions
+
+- The app may request access to your device's file system to save images you create. This is only used to store your drawings locally.
+
+## Analytics
+
+- The app does **not** use analytics or tracking services.
+
+## Changes
+
+- If this policy changes, updates will be posted in this document.
+
+## Contact
+
+If you have any questions about this privacy policy, please contact the developer.
+
