+++ capacitor.config.ts
@@ -1,8 +1,8 @@
 import type { CapacitorConfig } from '@capacitor/cli';
 
 const config: CapacitorConfig = {
-  appId: 'ayush.muley.painting_ai',
-  appName: 'painting-ai',
+  appId: 'ayush.muley.paintingai',
+  appName: 'painting',
   webDir: 'dist'
 };
