+++ index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <title>painting-ai</title>
+    <title>painting</title>
     
     <base href="/" />
     
@@ -20,7 +20,7 @@
     
     <!-- add to homescreen for ios -->
     <meta name="mobile-web-app-capable" content="yes" />
-    <meta name="apple-mobile-web-app-title" content="painting-ai" />
+    <meta name="apple-mobile-web-app-title" content="painting" />
     <meta name="apple-mobile-web-app-status-bar-style" content="black" />
   </head>
   <body>
