+++ ionic.config.json
@@ -1,5 +1,5 @@
 {
-  "name": "painting-ai",
+  "name": "painting",
   "integrations": {
     "capacitor": {}
   },
