+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "painting-ai",
+  "name": "painting",
   "private": true,
   "version": "0.0.1",
   "type": "module",
@@ -12,19 +12,19 @@
     "lint": "eslint"
   },
   "dependencies": {
-    "@capacitor/android": "^7.2.0",
+    "@capacitor/android": "7.2.0",
     "@capacitor/app": "7.0.1",
     "@capacitor/core": "7.2.0",
+    "@capacitor/filesystem": "^7.0.1",
     "@capacitor/haptics": "7.0.1",
     "@capacitor/keyboard": "7.0.1",
+    "@capacitor/share": "^7.0.1",
     "@capacitor/status-bar": "7.0.1",
-    "@genkit-ai/googleai": "^1.10.0",
+    "@capacitor/toast": "^7.0.1",
     "@ionic/react": "^8.5.0",
     "@ionic/react-router": "^8.5.0",
     "@types/react-router": "^5.1.20",
     "@types/react-router-dom": "^5.3.3",
-    "dotenv": "^16.5.0",
-    "genkit": "^1.10.0",
     "ionicons": "^7.4.0",
     "react": "19.0.0",
     "react-dom": "19.0.0",
