+++ public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "painting-ai",
-  "name": "painting-ai",
+  "short_name": "painting",
+  "name": "painting",
   "icons": [
     {
       "src": "assets/icon/favicon.png",
